+++ test/hmacKey.ts
@@ -42,10 +42,6 @@ const metadataResponse = {
   timeCreated: '20190101T00:00:00Z',
   updated: '20190101T00:00:00Z',
 };
-const hmacKeyResource = {
-  metadata: metadataResponse,
-  secret: 'secret-key',
-};
 
 describe('HmacKey', () => {
   beforeEach(() => {
@@ -108,184 +104,6 @@ describe('HmacKey', () => {
       assert.throws(() => {
         const _hmacKey = new HmacKey(STORAGE);
       }, /access ID is needed/);
-    });
-  });
-
-  describe('statics', () => {
-    // tslint:disable-next-line: no-any
-    let storageRequestStub: sinon.SinonStub<any, any>;
-    let STORAGE: Storage;
-
-    describe('createHmacKey', () => {
-      beforeEach(() => {
-        STORAGE = new Storage();
-        storageRequestStub = sandbox
-          .stub(STORAGE, 'request')
-          .callsFake((_opts: {}, callback: Function) => {
-            callback(null, hmacKeyResource);
-          });
-      });
-
-      it('should fail if serviceAccountEmail is not provided', () => {
-        return assert.rejects(
-          // tslint:disable-next-line: no-any
-          () => (STORAGE.createHmacKey as any)(undefined),
-          /first argument must be a service account email/
-        );
-      });
-
-      it('should return a promise when a callback is not provided', () => {
-        const promise = STORAGE.createHmacKey(SERVICE_ACCOUNT_EMAIL);
-        assert(promise instanceof Promise);
-        return promise;
-      });
-
-      it('should invoke callback with an HmacKeyResource', done => {
-        STORAGE.createHmacKey(
-          SERVICE_ACCOUNT_EMAIL,
-          (err: Error | null, hmacKey?: HmacKeyResource | null) => {
-            assert.ifError(err);
-            assert(hmacKey!.secret, hmacKeyResource.secret);
-
-            const hmacKeyInstance = hmacKey!.hmacKey;
-            assert(hmacKeyInstance instanceof HmacKey);
-            assert.strictEqual(
-              hmacKeyInstance.accessId,
-              hmacKeyResource.metadata.accessId
-            );
-            assert.deepStrictEqual(
-              hmacKeyInstance.metadata,
-              hmacKeyResource.metadata
-            );
-            done();
-          }
-        );
-      });
-    });
-
-    describe('getHmacKeys', () => {
-      beforeEach(() => {
-        STORAGE = new Storage();
-        storageRequestStub = sandbox
-          .stub(STORAGE, 'request')
-          .callsFake((_opts: {}, callback: Function) => {
-            callback(null, {});
-          });
-      });
-
-      it('should get HmacKeys without a query', done => {
-        STORAGE.getHmacKeys(() => {
-          const firstArg = storageRequestStub.firstCall.args[0];
-          assert.strictEqual(
-            firstArg.uri,
-            `/projects/${STORAGE.projectId}/hmacKeys`
-          );
-          assert.deepStrictEqual(firstArg.qs, {});
-          done();
-        });
-      });
-
-      it('should get HmacKeys with a query', done => {
-        const query = {
-          maxResults: 5,
-          pageToken: 'next-page-token',
-          serviceAccountEmail: SERVICE_ACCOUNT_EMAIL,
-          showDeletedKeys: false,
-        };
-
-        STORAGE.getHmacKeys(query, () => {
-          const firstArg = storageRequestStub.firstCall.args[0];
-          assert.strictEqual(
-            firstArg.uri,
-            `/projects/${STORAGE.projectId}/hmacKeys`
-          );
-          assert.deepStrictEqual(firstArg.qs, query);
-          done();
-        });
-      });
-
-      it('should execute callback with error', done => {
-        const error = new Error('Error.');
-        const apiResponse = {};
-
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(error, apiResponse);
-        });
-
-        STORAGE.getHmacKeys({}, (err, hmacKeys, nextQuery, resp) => {
-          assert.strictEqual(err, error);
-          assert.strictEqual(hmacKeys, null);
-          assert.strictEqual(nextQuery, null);
-          assert.strictEqual(resp, apiResponse);
-          done();
-        });
-      });
-
-      it('should return nextQuery if more results exist', done => {
-        const token = 'next-page-token';
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(null, {nextPageToken: token, items: []});
-        });
-
-        STORAGE.getHmacKeys({maxResults: 5}, (err, _hmacKeys, nextQuery) => {
-          assert.ifError(err);
-          // tslint:disable-next-line: no-any
-          assert.strictEqual((nextQuery as any).pageToken, token);
-          // tslint:disable-next-line: no-any
-          assert.strictEqual((nextQuery as any).maxResults, 5);
-          done();
-        });
-      });
-
-      it('should return null nextQuery if there are no more results', done => {
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(null, {items: []});
-        });
-
-        STORAGE.getHmacKeys({maxResults: 5}, (err, _hmacKeys, nextQuery) => {
-          assert.ifError(err);
-          assert.strictEqual(nextQuery, null);
-          done();
-        });
-      });
-
-      it('should return HmacKey objects', done => {
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(null, {items: [metadataResponse]});
-        });
-
-        STORAGE.getHmacKeys((err, hmacKeys) => {
-          assert.ifError(err);
-          assert(hmacKeys![0] instanceof HmacKey);
-          done();
-        });
-      });
-
-      it('should return apiResponse', done => {
-        const resp = {items: [metadataResponse]};
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(null, resp);
-        });
-
-        STORAGE.getHmacKeys((err, _hmacKeys, _nextQuery, apiResponse) => {
-          assert.ifError(err);
-          assert.deepStrictEqual(resp, apiResponse);
-          done();
-        });
-      });
-
-      it('should populate returned HmacKey object with accessId and metadata', done => {
-        storageRequestStub.callsFake((_opts: {}, callback: Function) => {
-          callback(null, {items: [metadataResponse]});
-        });
-
-        STORAGE.getHmacKeys((err, hmacKeys) => {
-          assert.ifError(err);
-          assert.strictEqual(hmacKeys![0].accessId, metadataResponse.accessId);
-          assert.deepStrictEqual(hmacKeys![0].metadata, metadataResponse);
-          done();
-        });
-      });
     });
   });
